refactor: add a postcard state to replace expanded
Adds a postcard state that will hold all of the current states of
the postcard contact form. Currently, the state of the card is set by
the boolean state `expanded` which only can hold two states. While this
change only uses two states, the change of type to a limited string
allows for additional states to be added in future commits