refactor:move css transition for flip card to TW
Removed the CSS classes from the global css to tailwind classes.
Previously these classes were only working in the css stylesheet because
Tailwind v3 didn't have support for 3D transitions. Now with the upgrade
to v4, it can all be placed in the same place, on the elements, for
easier readability.